+++ src/utils/BarNode.js
@@ -7,10 +7,12 @@ module.exports = class BarNode {
     this.ctx = ctx;
     this.data = data;
     this.color = "blue";
+
     return this;
   }
 
-  draw = () => {
+  draw = (color) => {
+    this.ctx.fillStyle = color;
     this.ctx.fillRect(this.x, this.y - this.height, this.width, this.height);
   };
 
@@ -44,8 +46,9 @@ module.exports = class BarNode {
       x < position.x + this.width &&
       y > position.y - this.height &&
       y < position.y
-    )
+    ) {
       return true;
+    }
 
     return false;
   };
